Using dynamic routing for people
