+++ app/people/page.tsx
@@ -3,23 +3,12 @@ import React, { useEffect, useState, useCallback, Fragment } from "react";
 import { useRouter, useSearchParams, usePathname } from 'next/navigation';
 import withApollo from "@/lib/withApollo";
 import { gql, useQuery, useLazyQuery } from "@apollo/client";
-import { ProjectType, UserType } from "../components/addAssignmentModal";
+import { UserType } from "../components/addAssignmentModal";
 import { DateTime, Interval } from "luxon";
 import { Listbox, Transition } from "@headlessui/react";
 import { CheckIcon, ChevronDownIcon } from "@heroicons/react/20/solid";
 import { WorkWeek, WorkWeekType } from "../components/workWeek";
-
-export const dynamic = 'force-dynamic'
-
-interface AssignmentType {
-	id: number;
-	endsOn: string | null;
-	startsOn: string | null;
-	status: string;
-	assignedUser: UserType;
-	workWeeks: [];
-	project: ProjectType;
-}
+import { parseProjectDates, parseWorkWeekDate, AssignmentType } from "../people/helperFunctions";
 
 const GET_USER_ASSIGNMENTS = gql`
 	query getUserAssignments($selectedUserId: ID!) {
@@ -65,7 +54,6 @@ const PeopleView: React.FC = () => {
 		name: "Select",
 	});
 	const router = useRouter();
-	const searchParams = useSearchParams();
 	const pathname = usePathname();
 
 	useEffect(() => {
@@ -86,114 +74,8 @@ const PeopleView: React.FC = () => {
 		errorPolicy: "all",
 	});
 
-	const [
-		getUserAssignments,
-		{
-			data: userAssignmentData,
-			loading: userAssignmentLoading,
-			error: userAssignmentError,
-			called,
-		},
-	] = useLazyQuery(GET_USER_ASSIGNMENTS, {
-		variables: { selectedUserId: selectedUser.id },
-	});
-
-	if (called && userAssignmentLoading) return <p>Loading User Assignments</p>;
-	if (userListLoading) return <p>Loading Users</p>;
-	if (userListError) return <p>Error Loading Users</p>;
-	if (userAssignmentError) return <p>Error Loading User Assignments</p>;
-
-	//convert a human readable date to calendar week and year
-	const parseProjectDates = (date: string) => {
-		return {
-			cweek: parseInt(DateTime.fromISO(date).toFormat("W")),
-			year: parseInt(DateTime.fromISO(date).toFormat("kkkk")),
-		};
-	};
-
-	//creates a human readable date from calendar week and year
-	const parseWorkWeekDate = (weekYear: number, weekNumber: number) => {
-		return DateTime.fromObject({ weekYear, weekNumber }).toISODate();
-	};
-
-	//an array of workWeeks with a formatted date added to the object
-	const workWeekArr = userAssignmentData?.userAssignments.flatMap(
-		(assignment: AssignmentType) =>
-			assignment.workWeeks.map((week: WorkWeekType) => {
-				const date = parseWorkWeekDate(week.year, week.cweek);
-				return { workWeek: week, date };
-			})
-	);
-
-	//an array of calendar dates, IF an assignment has a start and end date we set the workWeeks to follow those dates, otherwise we generate dates for current week thru end of year
-	const calWeekDatesArr = (
-		startDate: string | null,
-		endDate: string | null
-	) => {
-		return Interval.fromDateTimes(
-			//we set the start date to monday to ensure it captures that week
-			startDate
-				? DateTime.fromISO(startDate).set({ weekday: 1 })
-				: DateTime.now().set({ weekday: 1 }),
-			//we set the end date to tuesday to ensure it captures the monday of that week
-			endDate
-				? DateTime.fromISO(endDate).set({ weekday: 2 })
-				: DateTime.now().endOf("year").set({ weekday: 2 })
-			// DateTime.now().endOf("year").set({ weekday: 1 })
-		)
-			.splitBy({ days: 1 })
-			.filter((dur: Interval<true>) => dur.start.weekday === 1)
-			.map((dur: Interval<true>) => dur.start.toISODate());
-	};
-
-	//compare work week data and array of weeks and generate components accordingly
-	const workWeekComponentsArr = (
-		assignmentStartDate: string | null,
-		assignmentEndDate: string | null,
-		calWeekDatesArr: (
-			assignmentStartDate: string | null,
-			assignmentEndDate: string | null
-		) => string[],
-		workWeekArr: { workWeek: WorkWeekType; date: string }[],
-		assignmentId: number
-	) => {
-		return calWeekDatesArr(assignmentStartDate, assignmentEndDate).map(
-			(date: string) => {
-				//if a workWeek entry exists for a date include it in the array of work week inputs
-				const existingWeek = workWeekArr?.find(
-					(week: { workWeek: WorkWeekType; date: string }) =>
-						week.date == date && week.workWeek.assignmentId == assignmentId
-				);
-				if (existingWeek) {
-					return (
-						<WorkWeek
-							key={existingWeek.date}
-							workWeek={existingWeek.workWeek}
-						/>
-					);
-				}
-				const { cweek, year } = parseProjectDates(date);
-				return <WorkWeek key={date} workWeek={{ assignmentId, cweek, year }} />;
-			}
-		);
-	};
-
-	const createQueryString = 
-		(name: string, value: string) => {
-			const params = new URLSearchParams(searchParams.toString())
-			params.set(name, value)
-
-			return params.toString()
-		}
-	
-
 	const handleUserChange = (user: UserType) => {
-
-		router.push(pathname + '/' + encodeURIComponent(user.name))
-		console.log('user', user)
-
-		setSelectedUser(user);
-		getUserAssignments({ variables: { selectedUserId: user.id } });
+		router.push(pathname + '/' + encodeURIComponent(user.name.toString()));
 	};
 
 	return (
@@ -277,61 +159,6 @@ const PeopleView: React.FC = () => {
 					</>
 				)}
 			</Listbox>
-
-			<div className="flex flex-col items-start">
-				{userAssignmentData ? (
-					userAssignmentData.userAssignments.map(
-						(assignment: AssignmentType) => (
-							<div key={assignment.id} className="flex">
-								<div>
-									<p className="text-xl underline">
-										Project Name: {assignment.project.name}
-									</p>
-									{assignment.project.startsOn ? (
-										<p>Project Start Date: {assignment.project.startsOn}</p>
-									) : (
-										""
-									)}
-									{assignment.project.endsOn ? (
-										<p>Project End Date: {assignment.project.endsOn}</p>
-									) : (
-										""
-									)}
-									<div className="p-3">
-										<span>Assignment Status:{assignment.status}</span>
-										<br />
-										<span>Assignment Duration</span>
-										<br />
-										<span>Starts On: {assignment.startsOn}</span>
-										<br />
-										<span>Ends On: {assignment.endsOn}</span>
-										<br />
-									</div>
-								</div>
-								{assignment.project.startsOn ? (
-									<p>
-										Project Work Week:
-										{parseProjectDates(assignment.project.startsOn).cweek}
-									</p>
-								) : (
-									""
-								)}
-								<div className="p-3 flex">
-									{workWeekComponentsArr(
-										assignment.startsOn,
-										assignment.endsOn,
-										calWeekDatesArr,
-										workWeekArr,
-										assignment.id
-									)}
-								</div>
-							</div>
-						)
-					)
-				) : (
-					<p>User has no Assignments</p>
-				)}
-			</div>
 		</div>
 	);
 };
